mkshrc: include busybox path on global path
Busybox is a tool that is used by many developers as they develop and
test their code. Currently it is necessary to add it to one's path in
order to use it, and this must be done each time the target boots. Add
this to the global path instead. This will not introduce unnecessary
dependencies on busybox, as it is still not installed by default.

Change-Id: I3c44de9fdb661fd4feb99569dd77690d8a2a1294